+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+### v3.0.0
+    Client: 仅对 php 的 Cookie 进行验证
+    Server: aspx/ashx/jsp/jspx 使用全局变量替代了 Session, 已不再依赖 Cookie (特别感谢 @c0ny1 的解决方案)
+
+### v2.6.0
+    Client: 增加 `--write-interval` 参数，可调整 FORWARD 请求间隔
+    Client: 重新调整 `--read-interval` 默认值为 300
+    Clinet: 增加 `--target` 参数，实现端口转发功能
+
+### v2.5.1
+    Client: 当 Session 过期时，自动追加 Cookie 继续运行
+    Server: 修复 jsp(x) 在部分 jdk 低版本中找不到 java.nio.ByteBuffer.clear() 方法问题
+
+### v2.5.0
+    Client: 检测 socks 连接已关闭，则自动结束会话，减少大量请求流量
+    Client: 优化了异常信息输出，已能捕获高并发产生的异常
+    Client: 文档新增 Mac OSX 环境的运行建议
+    Client: 调整默认设置，流量减少约 46%
+    Client: 修复 `--file` 读取复杂文件编码转义问题
+
+### v2.4.1
+    添加 Session 过期提示
+    调整 `askGeorg` 检测请求，timeout 为 10 秒
+
+### v2.4.0
+    修复非 apache 的环境 BUG [php]
+
+### v2.3.2
+    修复 `--local-dns` 参数注释
+
 ### v2.3.1
     修复 python3 中 response 错误信息提醒 BUG
     优化了错误信息提醒
@@ -29,7 +59,7 @@
     优化了错误信息输出
     修改了目录名称 scripts/ => templates/ 和 neoreg_server/ => neoreg_servers/
     移除 socks4 的支持
-	移除 javascript tunnel 支持
+    移除 javascript tunnel 支持
 
 ### v1.5.0
     修复 php >= 7.1 版本，无法正常使用的问题